register & login styling completed & other pages created
