+++ app/arbswap/@auth/(.)forgot-password/page.tsx
@@ -1,0 +1,14 @@
+import Modal from '@/components/CommonComponents/Modal'
+import React from 'react'
+// import ForgotPassword from '@component/ForgotPassword'
+
+function ForgotPassword() {
+  return (
+    <Modal>
+      <></>
+      {/* <ForgotPassword /> */}
+    </Modal>
+  )
+}
+
+export default ForgotPassword
